Fix -o json with single cluster info
In case there is a `--cluster-name cluster`
specified for `cluster get`, it defaults to a
table view regardless of what the user specified.

Making it such that the display is detailed only if
the user wants a table.

Also fixing a bug where detailed display fails if the
cluster is not ready because the call to fetch the
allow lists fails.